feat: move lakehouse source to io directory
