+++ daft/io/delta_lake/delta_lake_scan.py
